requirements.txt
astroid==2.15.2
black==23.3.0
click==8.1.3
colorama==0.4.6
dill==0.3.6
grpcio==1.53.0
grpcio-tools==1.53.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
paho-mqtt==1.6.1
pathspec==0.11.1
platformdirs==3.2.0
protobuf==4.22.1
pylint==2.17.2
tomlkit==0.11.7
wrapt==1.15.0